backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
requests==2.32.3
openai==1.35.7
google-generativeai==0.7.2
sentence-transformers==3.0.1
scikit-learn==1.5.0
numpy==1.26.4
pydantic==2.7.4
pytest==8.2.2
pytest-asyncio==0.23.7
